move section stuff out of index
